What I did: -Added some listeners and handlers for widgets that came with the registration page in the last merge. -Removed the gearIcon and userIcon ImageViews from the landing page. There is no reason a user should see these before a login has occurred. Leave this as is.
We REALLY need to get this app finished in the next few days. If this means 14 hour days with no sleep, so be it!
------------------------------------------------------------------------------------------
Some issues for Mireya to fix:
-The userIcon(goes to profile) and the gearIcon(will go to a pin change page)
are ImageViews and these cannot be clicked. When I attempted to use Glide to place
the gear image and user image inside a button, it did not work. Glide should automatically re-sizes the image to fit the widget so I'm not sure what is happening here. Please fix this promptly so I can add functionality to these widgets.I tried
adding a transparent button over the top of these icons but it is moved out of the way
every time I tried and I don't have time to look up how to fix that.

-Widget constraints are still not working as intended. Please drag a constraint from each side of the widget to each of the four edges of the window. All widgets are inside this top level container and must be constrained to its edges unless you want to add child containers for them to go in.

-Fix all warnings and missing stuff in the XML

*** I need these changes to be done in the next 48 hours. I have many days of development left and I can't proceed with that until the GUI is fully functional. Time spent on the GUI is time I can't spend on critical backend stuff that still isn't fully operational. This does not even include the testing with the hardware. I have taken a cursory look into this stuff so ask me any questions if you need to. Thank you.

Please make sure this is fully functional when you test in the emulator!!!!
***

--------------------------------------------------------------------------
To implement Glide, please paste make sure the following dependencies are
in the build.gradle file(located near the bottom of the project pane):

implementation 'com.github.bumptech.glide:glide:4.7.1'
annotationProcessor 'com.github.bumptech.glide:compiler:4.7.1'

Glide documentation: https://bumptech.github.io/glide/

--------------------------------------------------------------------------------------
To be implemented in the next day or two:
1) SSL for the web server so we can use https

2) Another asynchronous task so for the registration page to send data to the database(currently I am manually entering data for testing).
How AsyncTask works and why we use it:
https://www.upwork.com/hiring/mobile/why-you-should-use-asynctask-in-android-development/

3) Biometric authentication(this will be painful and time consuming).

4) Another PHP restAPI to handle the registration process.
-----------------------------------------------------------------------------------------
As this is the most functional version of our software, I will me merging into the master branch. Please pull from master to keep current.
-----------------------------------------------------------------------------------------